repo restructing and cleaning to support socketio
