+++ misc/tracing/tracing-server.js
@@ -1,85 +1,109 @@
-const fs = require('fs');
-const path = require('path');
 const express = require('express');
+const socketio = require('socket.io');
+const cors = require('cors');
+const { NodeSDK } = require('@opentelemetry/sdk-node');
+const { OTLPTraceExporter } = require('@opentelemetry/exporter-trace-otlp-http');
+const { HttpInstrumentation } = require("@opentelemetry/instrumentation-http");
+const { PgInstrumentation } = require("@opentelemetry/instrumentation-pg");
+
+//Open Telemetry Auto Instrumentation
+const sdk = new NodeSDK({
+  serviceName: 'invest-with-friends',
+  traceExporter: new OTLPTraceExporter({url:'http://localhost:44222/v1/traces'}),
+  instrumentations: [new HttpInstrumentation(), new PgInstrumentation()]
+});
+sdk.start();
+
+// Express Server
 const app = express();
-const port = 3001;
-let cors = require('cors');
-const { NodeTracerProvider } = require('@opentelemetry/sdk-trace-node');
-const { registerInstrumentations } = require('@opentelemetry/instrumentation');
-const { ConsoleSpanExporter, BatchSpanProcessor } = require('@opentelemetry/sdk-trace-base');
-const { HttpInstrumentation } = require('@opentelemetry/instrumentation-http');
-const { PgInstrumentation } = require('@opentelemetry/instrumentation-pg');
+const PORT = 44222;
 
-// Optionally register instrumentation libraries
-registerInstrumentations({
-	instrumentations: [new HttpInstrumentation(), new PgInstrumentation()],
-});
-
-// Create a tracer provider
-const provider = new NodeTracerProvider();
-
-// Create a span exporter and exports in console in batches
-const exporter = new ConsoleSpanExporter();
-const processor = new BatchSpanProcessor(exporter);
-provider.addSpanProcessor(processor);
-
-// Initialize the tracer provider
-provider.register();
-
-// Log data to .txt file
-const rootDirectory = path.resolve(__dirname, '..');
-const logFilePath = path.join(rootDirectory, 'log.txt');
-
-function logData(data) {
-	fs.appendFileSync(logFilePath, data + '\n');
-}
-
-// Add a span processor to log the span data
-provider.addSpanProcessor({
-	// onStart needed or run into error saying onStart is not a function
-	onStart: (span) => {
-		// Doing nothing on start
-	},
-	onEnd: (span) => {
-		if (span.attributes['http.target'] === '/api/log-xhr-data' && span.attributes['http.method'] === 'POST') {
-			return;
-		}
-		if (span.attributes['http.method'] !== 'OPTIONS') {
-			if (span.attributes['http.method']) {
-				logData(`HTTP method: ${span.attributes['http.method']}`);
-			}
-			if (span.attributes['http.target']) {
-				logData(`API route: ${span.attributes['http.target']}`);
-			}
-			if (span.attributes['http.method'] && span.duration) {
-				logData(`Execution time (ms): ${span.duration}`);
-			}
-			if (span.attributes['http.status_code']) {
-				logData(`HTTP status code: ${span.attributes['http.status_code']}`);
-			}
-			if (span.attributes['db.statement']) {
-				logData(`SQL Query: ${span.attributes['db.statement']}`);
-			}
-		}
-	},
-});
-
-// Express logic below
-app.use(
-	cors({
-		credentials: true,
-	})
-);
-
+app.use(cors({credentials: true,}));
 app.use(express.json());
 
-app.post('/api/log-xhr-data', (req, res) => {
-	const { requestPayload, responseData } = req.body;
-	logData(`Request Payload: ${requestPayload}`);
-	logData(`Response Data: ${responseData}`);
-	res.json({ status: 'success', message: 'XHR data received' });
+const expressServer = app.listen(PORT, () => console.log(`Injected server listening on port: ${PORT}...`));
+
+//Socket IO server
+const io = socketio(expressServer);
+let socketId;
+
+let stackData = {};
+
+app.post('/v1/traces', (req, res) => {
+  filter(req.body);
+  if (Object.keys(stackData).length >= 7) {
+    io.to(socketId).emit('interaction', {data: stackData});
+    console.log('sending data...');
+    stackData = {};
+  }
+  // io.to(socketId).emit('interaction', {data: req.body});
+  res.sendStatus(200);
+})
+
+io.on('connection', (socket) => {
+
+  socket.on('socketId', (data) => {
+    socketId = data.data;
+    console.log(`connected to new socket: id ${socketId}`);
+  });
+
+  socket.on('disconnect', () => {
+    console.log(`id ${socketId} disconnected`)
+  })
 });
 
-app.listen(port, () => {
-	console.log(`Server listening on port ${port}`);
-});
+//request filtering logic
+const filter = (requestBody) => {
+
+  const keys = new Set([
+    'requestPayload',
+    'responseData',
+  ])
+
+  //check if open telemetry request
+  if (requestBody.resourceSpans) {
+
+    const scopeSpans = requestBody.resourceSpans[0].scopeSpans;
+      
+    for (const scopeSpan of scopeSpans) {
+      if (scopeSpan.scope.name === '@opentelemetry/instrumentation-pg') {
+        const spans = scopeSpan.spans;
+        for (const span of spans) {
+
+          //crawl attributes for sql query
+          for (const obj of span.attributes) {
+            if (obj.key === 'db.statement') {
+              stackData['sqlQuery'] = obj.value.stringValue
+            }
+          }
+        }
+      }
+      else if (scopeSpan.scope.name === '@opentelemetry/instrumentation-http') {
+        const spans = scopeSpan.spans;
+        for (const span of spans) {
+
+          //filter out noise
+          const url = span.attributes[0].value.stringValue;
+          if (url.includes('/socket.io/') || url.includes('/v1/traces')) {continue;} 
+          if (span.name === 'OPTIONS') {continue}
+
+          //crawl remaining spans for relevant properties
+          else {
+            stackData['executionTime'] = `${(span.endTimeUnixNano - span.startTimeUnixNano)/1000000000}s`
+            for (const attribute of span.attributes) {
+              if (attribute.key === 'http.url') {stackData['route'] = attribute.value.stringValue}
+              if (attribute.key === 'http.method') {stackData['httpMethod'] = attribute.value.stringValue}
+              if (attribute.key === 'http.status_code') {stackData['statusCode'] = attribute.value.intValue}
+            }
+          }
+        }
+      }
+    }
+  } else {
+    for (const [k, v] of Object.entries(requestBody)) {
+      if (keys.has(k)) {
+        stackData[k] = v;
+      }
+    }
+  }
+}
